format

+++ src/relax/distributed/transform/propagate_sharding.cc
@@ -56,12 +56,17 @@ void CollectAxisGraphBinary(const VarBindingNode* binding, const CallNode* call,
 void CollectAxisGraphUnary(const VarBindingNode* binding, const CallNode* call,
                            AxisGroupGraph* axis_group_graph) {
   const std::vector<std::string> unary_op_names = {
-      "abs",   "acos",     "acosh",    "asin",   "asinh", "atan",
-      "atanh", "ceil",     "cos",      "cosh",   "exp",   "floor",
-      "log",   "negative", "nn.relu",  "round",  "rsqrt", "sigmoid",
-      "sign",  "sin",      "sinh",     "square", "sqrt",  "tan",
-      "tanh",  "clip",     "isfinite", "isinf",  "isnan", "dist.annotate_sharding",
-      "erf",   "nn.gelu", "builtin.stop_lift_params"};
+      "abs",    "acos",     "acosh",
+      "asin",   "asinh",    "atan",
+      "atanh",  "ceil",     "cos",
+      "cosh",   "exp",      "floor",
+      "log",    "negative", "nn.relu",
+      "round",  "rsqrt",    "sigmoid",
+      "sign",   "sin",      "sinh",
+      "square", "sqrt",     "tan",
+      "tanh",   "clip",     "isfinite",
+      "isinf",  "isnan",    "dist.annotate_sharding",
+      "erf",    "nn.gelu",  "builtin.stop_lift_params"};
   for (const auto& op_name : unary_op_names) {
     const Op& unary_op = Op::Get("relax." + op_name);
     if (call->op.same_as(unary_op)) {
@@ -162,30 +167,30 @@ class AxisGroupGraphBuilder : public ExprVisitor {
 
   void VisitBinding_(const VarBindingNode* binding, const TupleGetItemNode* val) {
     axis_group_graph_->JoinAxis(Axis(val->tuple.get(), -1, val->index), {binding->var.get(), -1},
-                               distributed::AxisGroupGraph::EdgeType::kDescend);
+                                distributed::AxisGroupGraph::EdgeType::kDescend);
     const auto* tensor_sinfo = GetStructInfoAs<TensorStructInfoNode>(binding->var);
-    if(!tensor_sinfo){
+    if (!tensor_sinfo) {
       ExprVisitor::VisitBinding_(binding, val);
       return;
     }
     int ndim = tensor_sinfo->ndim;
     for (int i = 0; i < ndim; i++) {
       axis_group_graph_->JoinAxis(Axis(val->tuple.get(), i, val->index), {binding->var.get(), i},
-                                 distributed::AxisGroupGraph::EdgeType::kDescend);
+                                  distributed::AxisGroupGraph::EdgeType::kDescend);
     }
     ExprVisitor::VisitBinding_(binding, val);
   }
 
   void VisitBinding_(const VarBindingNode* binding, const VarNode* val) {
     Array<TensorStructInfo> tensor_sinfos;
-    if(const auto* tensor_sinfo = binding->var->struct_info_.as<TensorStructInfoNode>()){
+    if (const auto* tensor_sinfo = binding->var->struct_info_.as<TensorStructInfoNode>()) {
       tensor_sinfos.push_back(GetRef<TensorStructInfo>(tensor_sinfo));
-    } else if(const auto* tuple_sinfo = binding->var->struct_info_.as<TupleStructInfoNode>()){
+    } else if (const auto* tuple_sinfo = binding->var->struct_info_.as<TupleStructInfoNode>()) {
       ICHECK(tuple_sinfo);
-      for(const auto& sinfo: tuple_sinfo->fields) {
+      for (const auto& sinfo : tuple_sinfo->fields) {
         tensor_sinfos.push_back(Downcast<TensorStructInfo>(sinfo));
       }
-    } else{
+    } else {
       ExprVisitor::VisitBinding_(binding, val);
       return;
     }
@@ -193,7 +198,7 @@ class AxisGroupGraphBuilder : public ExprVisitor {
       int ndim = tensor_sinfos[idx]->ndim;
       for (int i = -1; i < ndim; i++) {
         axis_group_graph_->JoinAxis({val, i, idx}, {binding->var.get(), i, idx},
-                                  distributed::AxisGroupGraph::EdgeType::kDescend);
+                                    distributed::AxisGroupGraph::EdgeType::kDescend);
       }
     }
     ExprVisitor::VisitBinding_(binding, val);
@@ -354,11 +359,12 @@ class DistributedIRBuilder : public ExprMutator {
  private:
   using ExprMutator::VisitExpr_;
 
-  DTensorStructInfo ConvertToDTensorStructInfo(TensorStructInfo sinfo, Expr expr,  int tuple_idx=0) {
+  DTensorStructInfo ConvertToDTensorStructInfo(TensorStructInfo sinfo, Expr expr,
+                                               int tuple_idx = 0) {
     int ndim = sinfo->ndim;
     DeviceMesh device_mesh =
         std::get<0>(axis_group_graph_.GetAxisShardingSpec({expr.get(), -1, tuple_idx})).first;
-    ICHECK(device_mesh.defined()) << expr << "["<<tuple_idx<<"] is not assigned device mesh";
+    ICHECK(device_mesh.defined()) << expr << "[" << tuple_idx << "] is not assigned device mesh";
     Array<PlacementSpec> placement_specs(
         std::vector<PlacementSpec>(device_mesh->shape.size(), PlacementSpec::Replica()));
     for (int i = 0; i < ndim; i++) {
@@ -371,21 +377,20 @@ class DistributedIRBuilder : public ExprMutator {
         placement_specs.Set(sharding_dim, PlacementSpec::Sharding(i));
       }
     }
-    return DTensorStructInfo(sinfo, device_mesh,
-                             Placement(placement_specs));
+    return DTensorStructInfo(sinfo, device_mesh, Placement(placement_specs));
   }
 
   Expr RewriteInputTensorAndConstant(Expr tensor) {
     StructInfo new_sinfo;
     if (tensor->struct_info_.as<TensorStructInfoNode>()) {
-      new_sinfo = ConvertToDTensorStructInfo(
-          Downcast<TensorStructInfo>(tensor->struct_info_), tensor);
-    } else if (const auto* tuple = tensor->struct_info_.as<TupleStructInfoNode>()){
+      new_sinfo =
+          ConvertToDTensorStructInfo(Downcast<TensorStructInfo>(tensor->struct_info_), tensor);
+    } else if (const auto* tuple = tensor->struct_info_.as<TupleStructInfoNode>()) {
       Array<StructInfo> tuple_sinfo_fields;
-      for (int i = 0; i < tuple->fields.size(); i++) {
-        if(tuple->fields[i].as<TensorStructInfoNode>()){
-          tuple_sinfo_fields.push_back(ConvertToDTensorStructInfo(
-              Downcast<TensorStructInfo>(tuple->fields[i]), tensor, i));
+      for (int i = 0; i < static_cast<int>(tuple->fields.size()); i++) {
+        if (tuple->fields[i].as<TensorStructInfoNode>()) {
+          tuple_sinfo_fields.push_back(
+              ConvertToDTensorStructInfo(Downcast<TensorStructInfo>(tuple->fields[i]), tensor, i));
         } else {
           tuple_sinfo_fields.push_back(tuple->fields[i]);
         }
@@ -447,10 +452,9 @@ class DistributedIRBuilder : public ExprMutator {
     ObjectPtr<CallNode> n = make_object<CallNode>(*new_call.get());
     if (new_call->op.same_as(call_tir_op)) {
       // do not infer output sinfo when arg size is 0
-      if(!args.empty()){
+      if (!args.empty()) {
         n->args.Set(1, Tuple(args));
-        n->sinfo_args = {InferShardingSpec(
-            Call(n), this->builder_, new_call->sinfo_args[0], f)};
+        n->sinfo_args = {InferShardingSpec(Call(n), this->builder_, new_call->sinfo_args[0], f)};
       }
     } else {
       n->args = args;
@@ -466,7 +470,7 @@ class DistributedIRBuilder : public ExprMutator {
     return Call(n);
   }
 
-  Expr RemoveAnnotateSharding(Call call){
+  Expr RemoveAnnotateSharding(Call call) {
     static const Op& annotate_sharding_op = Op::Get("relax.dist.annotate_sharding");
     if (call->op.same_as(annotate_sharding_op)) {
       return call->args[0];
@@ -475,39 +479,37 @@ class DistributedIRBuilder : public ExprMutator {
     }
   }
 
-
-  Expr InsertRedistribute(Expr expr, DeviceMesh device_mesh, Placement placement){
+  Expr InsertRedistribute(Expr expr, DeviceMesh device_mesh, Placement placement) {
     return redistribute(expr, device_mesh, placement);
   }
 
-  Call RewriteOutSinfo(Call call, DeviceMesh device_mesh, Array<Placement> placements){
+  Call RewriteOutSinfo(Call call, DeviceMesh device_mesh, Array<Placement> placements) {
     // in cases when infer fails (like arg size is 0), we use propagated sinfo for output
     Call new_call = call;
     static Op call_tir_op = Op::Get("relax.call_tir");
     if (const auto* extern_func = call->op.as<ExternFuncNode>()) {
       if (extern_func->global_symbol == "vm.builtin.distributed.attention_kv_cache_view") {
         ObjectPtr<CallNode> new_call_node = make_object<CallNode>(*call.get());
-        StructInfo new_dtensor_sinfo =
-            DTensorStructInfo(Downcast<TensorStructInfo>(call->sinfo_args[0]), device_mesh,
-                              placements[0]);
+        StructInfo new_dtensor_sinfo = DTensorStructInfo(
+            Downcast<TensorStructInfo>(call->sinfo_args[0]), device_mesh, placements[0]);
         new_call_node->sinfo_args = {new_dtensor_sinfo};
         new_call = Call(new_call_node);
         new_call->struct_info_ = new_dtensor_sinfo;
       }
-    } else if (call->op.same_as(call_tir_op)){
+    } else if (call->op.same_as(call_tir_op)) {
       ICHECK(call->sinfo_args.size() == 1);
       if (!SinfoCompatibleWithDistIR(call->sinfo_args)) {
         ObjectPtr<CallNode> new_call_node = make_object<CallNode>(*call.get());
-        if(placements.size()==1){
-          new_call_node->sinfo_args = {DTensorStructInfo(Downcast<TensorStructInfo>(call->sinfo_args[0]), device_mesh,
-                              placements[0])};
+        if (placements.size() == 1) {
+          new_call_node->sinfo_args = {DTensorStructInfo(
+              Downcast<TensorStructInfo>(call->sinfo_args[0]), device_mesh, placements[0])};
         } else {
           const auto* tuple_sinfo = call->sinfo_args[0].as<TupleStructInfoNode>();
           ICHECK(placements.size() == tuple_sinfo->fields.size());
           Array<StructInfo> new_tuple_sinfo_fields;
-          for (int i = 0; i < static_cast<int>(placements.size()); i++){
-            new_tuple_sinfo_fields.push_back(DTensorStructInfo(Downcast<TensorStructInfo>(tuple_sinfo->fields[i]), device_mesh,
-                              placements[i]));
+          for (int i = 0; i < static_cast<int>(placements.size()); i++) {
+            new_tuple_sinfo_fields.push_back(DTensorStructInfo(
+                Downcast<TensorStructInfo>(tuple_sinfo->fields[i]), device_mesh, placements[i]));
           }
           new_call_node->sinfo_args = {TupleStructInfo(new_tuple_sinfo_fields)};
         }
@@ -520,13 +522,13 @@ class DistributedIRBuilder : public ExprMutator {
 
   void VisitBinding_(const VarBindingNode* binding, const CallNode* val) {
     Array<TensorStructInfo> orig_output_tensor_sinfos;
-    if(const auto* tensor_sinfo = GetStructInfoAs<TensorStructInfoNode>(binding->var)){
+    if (const auto* tensor_sinfo = GetStructInfoAs<TensorStructInfoNode>(binding->var)) {
       orig_output_tensor_sinfos.push_back(GetRef<TensorStructInfo>(tensor_sinfo));
-    } else if(const auto* tuple_sinfo = GetStructInfoAs<TupleStructInfoNode>(binding->var)){
-      for(const auto& sinfo: tuple_sinfo->fields) {
+    } else if (const auto* tuple_sinfo = GetStructInfoAs<TupleStructInfoNode>(binding->var)) {
+      for (const auto& sinfo : tuple_sinfo->fields) {
         orig_output_tensor_sinfos.push_back(Downcast<TensorStructInfo>(sinfo));
       }
-    } else{
+    } else {
       ExprMutator::VisitBinding_(binding, val);
       return;
     }
@@ -535,7 +537,7 @@ class DistributedIRBuilder : public ExprMutator {
         std::get<0>(axis_group_graph_.GetAxisShardingSpec({binding->var.get(), -1})).first;
     ICHECK(device_mesh.defined());
     Array<Placement> placements;  // every tuple element has a placement
-    for (int idx = 0;idx<static_cast<int>(orig_output_tensor_sinfos.size());idx++){
+    for (int idx = 0; idx < static_cast<int>(orig_output_tensor_sinfos.size()); idx++) {
       Array<PlacementSpec> placement_specs(
           std::vector<PlacementSpec>(device_mesh->shape.size(), PlacementSpec::Replica()));
       for (int i = 0; i < orig_output_tensor_sinfos[idx]->ndim; i++) {
@@ -554,29 +556,32 @@ class DistributedIRBuilder : public ExprMutator {
     new_call =
         Downcast<Call>(builder_->Normalize(RewriteOutSinfo(new_call, device_mesh, placements)));
 
-    if(const auto* inferred_dtensor_sinfo = new_call->struct_info_.as<DTensorStructInfoNode>()){
+    if (const auto* inferred_dtensor_sinfo = new_call->struct_info_.as<DTensorStructInfoNode>()) {
       Expr new_value = RemoveAnnotateSharding(new_call);
-      if (!StructuralEqual()(DTensorStructInfo(inferred_dtensor_sinfo->tensor_sinfo, device_mesh,
-                                          placements[0]),
-                          new_call->struct_info_)) {
+      if (!StructuralEqual()(
+              DTensorStructInfo(inferred_dtensor_sinfo->tensor_sinfo, device_mesh, placements[0]),
+              new_call->struct_info_)) {
         new_value = InsertRedistribute(new_value, device_mesh, placements[0]);
       }
-      if(const auto* var = new_value.as<VarNode>()){
+      if (const auto* var = new_value.as<VarNode>()) {
         var_remap_[binding->var->vid] = GetRef<Var>(var);
       } else {
         ReEmitBinding(binding, builder_->Normalize(new_value));
       }
-    } else{
+    } else {
       const auto* inferred_tuple_sinfo = new_call->struct_info_.as<TupleStructInfoNode>();
       ICHECK(inferred_tuple_sinfo) << new_call;
       Var new_var = builder_->Emit(new_call);
-      var_remap_[binding->var->vid]= new_var;
-      for (int i = 0; i < inferred_tuple_sinfo->fields.size(); i++) {
-        if (!StructuralEqual()(DTensorStructInfo(Downcast<DTensorStructInfo>(inferred_tuple_sinfo->fields[i])->tensor_sinfo, device_mesh,
-                                            placements[i]),
-                            inferred_tuple_sinfo->fields[i])) {
-          Var redistribute_var = builder_->Emit(InsertRedistribute(TupleGetItem(new_var, i), device_mesh, placements[i]));
-          tuple_getitem_remap_[TupleGetItem(binding->var, i)]=  redistribute_var;
+      var_remap_[binding->var->vid] = new_var;
+      for (int i = 0; i < static_cast<int>(inferred_tuple_sinfo->fields.size()); i++) {
+        if (!StructuralEqual()(
+                DTensorStructInfo(
+                    Downcast<DTensorStructInfo>(inferred_tuple_sinfo->fields[i])->tensor_sinfo,
+                    device_mesh, placements[i]),
+                inferred_tuple_sinfo->fields[i])) {
+          Var redistribute_var = builder_->Emit(
+              InsertRedistribute(TupleGetItem(new_var, i), device_mesh, placements[i]));
+          tuple_getitem_remap_[TupleGetItem(binding->var, i)] = redistribute_var;
         }
       }
     }
